+++ lib/common/widgets/custom_shapes/containers/circular_container.dart
@@ -1,0 +1,36 @@
+import 'package:flutter/material.dart';
+
+import '../../../../utils/constants/colors.dart';
+
+class CustomCircularContainer extends StatelessWidget {
+  const CustomCircularContainer({
+    super.key,
+    this.child,
+    this.width = 400,
+    this.height = 400,
+    this.radius = 400,
+    this.padding = 0,
+    this.backgroundColor = CustomColors.textWhite,
+  });
+
+  final double? width;
+  final double? height;
+  final double radius;
+  final double padding;
+  final Widget? child;
+  final Color backgroundColor;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      width: width,
+      height: height,
+      padding: EdgeInsets.all(padding),
+      decoration: BoxDecoration(
+        borderRadius: BorderRadius.circular(radius),
+        color: backgroundColor,
+      ),
+      child: child,
+    );
+  }
+}
