+++ lib/features/shop/screens/home/home.dart
@@ -1,6 +1,8 @@
 import 'package:ecommerce_app/utils/constants/colors.dart';
 import 'package:flutter/material.dart';
 
+import '../../../../common/widgets/custom_shapes/containers/circular_container.dart';
+
 class HomeScreen extends StatelessWidget {
   const HomeScreen({super.key});
 
@@ -15,55 +17,14 @@ class HomeScreen extends StatelessWidget {
               padding: EdgeInsets.all(0),
               child: Stack(
                 children: [
-                  CustomCircularContainer(),
-                  Container(
-                    width: 400,
-                    height: 400,
-                    padding: EdgeInsets.all(0),
-                    decoration: BoxDecoration(
-                      borderRadius: BorderRadius.circular(400),
-                      color: CustomColors.textWhite.withOpacity(0.1),
-                    ),
-                  )
+                  CustomCircularContainer(backgroundColor: CustomColors.textWhite.withOpacity(0.1),),
+                  CustomCircularContainer(backgroundColor: CustomColors.textWhite.withOpacity(0.1),),
                 ],
               ),
             )
           ],
         ),
       )
-    );
-  }
-}
-
-class CustomCircularContainer extends StatelessWidget {
-  const CustomCircularContainer({
-    super.key,
-    this.child,
-    this.width = 400,
-    this.height = 400,
-    this.radius = 400,
-    this.padding = 0,
-    this.backgroundColor = CustomColors.textWhite,
-  });
-
-  final double? width;
-  final double? height;
-  final double radius;
-  final double padding;
-  final Widget? child;
-  final Color backgroundColor;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: 400,
-      height: 400,
-      padding: EdgeInsets.all(0),
-      decoration: BoxDecoration(
-        borderRadius: BorderRadius.circular(400),
-        color: CustomColors.textWhite.withOpacity(0.1),
-      ),
-      child: child,
     );
   }
 }
